requirements.txt
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
black==23.9.1
click==8.1.7
decouple==0.0.7
exceptiongroup==1.1.3
fancycompleter==0.9.1
fastapi==0.103.2
flake8==6.1.0
h11==0.14.0
idna==3.4
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
pdbpp==0.10.3
platformdirs==3.10.0
pycodestyle==2.11.0
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
Pygments==2.16.1
pyrepl==0.9.0
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.23.2
wmctrl==0.5
